requirements.txt
Django==3.1
djangorestframework==3.11.1
psycopg2-binary==2.8.5
python-dotenv==0.14.0
requests==2.24.0
requests-cache==0.5.2
gunicorn==20.0.4
django-heroku==0.3.1
responses==0.11.0
